wip5

+++ views/_Nomenclature/Models/UsersModel.php
@@ -3,13 +3,21 @@
 namespace Views\_Nomenclature\Models;
 
 use Views\_AddEdit\Models\Handler;
+use Views\_Globals\Models\User;
 
 class UsersModel extends Handler
 {
-
+    /**
+    * Errors
+    */
+    const LOG_PASS_EMPTY = 344;
     const NAME_EMPTY = 345;
     const NO_SUCH_USER = 346;
     const INSERT_UPDATE_FAIL = 347;
+    const PERMISSION_DENIED = 348;
+
+    const USER_ADD_SUCCESS = 349;
+    const PERMISSIONS_ADD_SUCCESS = 350;
 
 
     public function __construct()
@@ -17,6 +25,15 @@ class UsersModel extends Handler
         parent::__construct();
         $this->connectDBLite();
     }
+
+    /**
+     * @throws \Exception
+     */
+    public function getAllPermissions()
+    {
+        return $this->findAsArray("SELECT id,name,description FROM permissions");
+    }
+
 
 
     /**
@@ -40,7 +57,6 @@ class UsersModel extends Handler
                         $uLocationsNames[$workingCenter['name']] .= $workingCenter['descr'] . ", ";
                     }
                 }
-
             }
             $user['locNames'] = $uLocationsNames;
         }
@@ -48,18 +64,44 @@ class UsersModel extends Handler
         return $users;
     }
 
+    public function addUserPermissions( array &$user ) : void
+    {
+        $allPermissions = $this->getAllPermissions();
+        $userPermissions = $this->findAsArray("SELECT permission_id as id FROM user_permissions WHERE user_id='{$user['id']}'");
+
+        $user['permissions'] = [];
+        foreach ($userPermissions as $permID)
+        {
+            foreach ($allPermissions as $permission)
+            {
+                if ( $permission['id'] == $permID['id'] ) $user['permissions'][$permID['id']] = $permission;
+            }
+        }
+    }
 
     /**
      * @param array $data
+     * Пост данные
      * @param bool $add
+     * Флаг добавления нового юзера
+     *
      * @return array
      * @throws \Exception
      */
     public function editUserData( array $data, bool $add = false ) : array
     {
+        // Проверка доступа для редактирования данных юзера
+        if ( !User::permission('nomUsers_edit') ) return ['error' => self::PERMISSION_DENIED ];
+
+        // А есть ли такой юзер?
         $userID = (int)$data['editUser_ID'];
         if ( !$add )
             if ( !$this->checkID( $userID,'users' ) ) return ['error' => self::NO_SUCH_USER ];
+        
+        // Берем старый доступ для проверки, может ли текущий изменять его данные
+        $access_old = (int)$this->findOne("SELECT access FROM users WHERE id='$userID'")['access'];
+        if ( $access_old === 1 && (User::getAccess() !== 1) ) return ['error' => self::PERMISSION_DENIED ];
+
 
         $userFirstName  = mysqli_escape_string($this->connection, htmlentities( trim($data['userFirstName']  ),  ENT_QUOTES) );
         if ( empty($userFirstName) ) return ['error' => self::NAME_EMPTY ];
@@ -70,11 +112,11 @@ class UsersModel extends Handler
         $fio = $userFirstName;
         if ( !empty($userSecondName) )
         {
-            $arrChars = preg_split('//u',$userSecondName,-1,PREG_SPLIT_NO_EMPTY);
+            $arrChars = preg_split('//u',$userSecondName, -1, PREG_SPLIT_NO_EMPTY);
             $fio .= " " . mb_strtoupper($arrChars[0],'UTF-8') . ".";
             if ( !empty($userThirdName) )
             {
-                $arrChars = preg_split('//u',$userThirdName,-1,PREG_SPLIT_NO_EMPTY);
+                $arrChars = preg_split('//u',$userThirdName, -1, PREG_SPLIT_NO_EMPTY);
                 $fio .=  mb_strtoupper($arrChars[0],'UTF-8') . ".";
             }
         }
@@ -84,10 +126,9 @@ class UsersModel extends Handler
         $userPass = mysqli_escape_string($this->connection, htmlentities( trim($data['userPass']  ),  ENT_QUOTES) );
 
         $location = implode(',', array_unique($data['wcList']??[]));
+        $presetID = $this->userRulesPreset( $data['userMTProd'] );
 
-        $userRulesPreset = $this->userRulesPreset($data['userMTProd']);
-
-        $row = [
+        $newUserRow = [
             [
                 'id'=>$userID ? $userID : '',
                 'login'=>$userLog,
@@ -95,14 +136,45 @@ class UsersModel extends Handler
                 'fio'=>$fio,
                 'fullFio'=>$fullFio,
                 'location'=>$location,
-                'access'=>$userRulesPreset
+                'access'=>$presetID
             ]
         ];
 
-        //debug($row,'ff',1);
-        if ( $lastID = $this->insertUpdateRows($row,'users') ) return ['success' => $lastID];
+        $result = [];
+        if ( $lastID = $this->insertUpdateRows($newUserRow,'users') )
+        {
+            $result = ['success' => self::USER_ADD_SUCCESS];
+        } else {
+            $result = ['error' => self::INSERT_UPDATE_FAIL ];
+        }
 
-        return ['error' => self::INSERT_UPDATE_FAIL ];
+        // Вносим данные в табл user_permissions
+        if ( $access_old != $presetID )
+        {
+            $permissionPresetOrigin = $this->permissionsPreset( $presetID );
+            $permArr = [
+                'id' => '',
+                'user_id' => !$add ? $userID : $lastID,
+                'permission_id' => '',
+                'date' => date('Y-m-d'),
+            ];
+            $permissionsPreset = [];
+            foreach ($permissionPresetOrigin as $value) 
+            {
+               $permArr['permission_id'] = $value;
+               $permissionsPreset[] = $permArr;
+            }
+            // удалим старые разрешения
+            if ( !$add ) $this->baseSql("DELETE FROM user_permissions WHERE user_id='$userID'");
+
+            if ( $this->insertUpdateRows($permissionsPreset,'user_permissions') ) {
+                $result = ['success' => self::PERMISSIONS_ADD_SUCCESS];
+            } else {
+                $result = ['error' => self::INSERT_UPDATE_FAIL ];
+            }
+        }
+
+        return $result;
     }
 
     /**
@@ -129,8 +201,56 @@ class UsersModel extends Handler
         }
         return $rulesPreset;
     }
+    protected function permissionsPreset( int $presetID ) : array
+    {
+        if ( $presetID === 1 )
+        {
+            $allPerm = $this->findAsArray("SELECT id FROM permissions");
+            foreach ($allPerm as $key => $value) 
+            {
+                $allPerm[$key] = $value['id'];
+            }
+        }
+
+        $result = [
+            // админ
+            1 => $allPerm ?? [],
+            // Moderator
+            122 => [
+                35,36,38,40,42
+            ],
+            // 3D modeller
+            2 => [
+                1,2,3,4,5,6,8,9,10,13,15,16,17,18,19,20,21,22,23,24,25,27,28,29,31,32,33,35,36,37,38,45
+            ],
+            // 3D Printing
+            3 => [
+                11,24,28,32,33,35,36,38,45
+            ],
+            // производство
+            4 => [
+                24,28,45
+            ],
+            // Модельер-доработчтк
+            5 => [
+                7,12,19,20,21,23,24,26,27,28,32,33,35,38,45
+            ],
+            // Технолог ЮВ (Валентин)
+            7 => [
+                28,32,33,36,38,45
+            ],
+            // ПДО
+            8 => [
+                3,7,10,13,21,22,23,24,27,28,35,36,38,45
+            ],
+        ];
+
+        if ( array_key_exists($presetID, $result) ) return $result[$presetID];
+        return [];
+    }
 
     /**
+     * Добавляем нового пользователя
      * @param $data
      * @return array
      * @throws \Exception
@@ -145,9 +265,16 @@ class UsersModel extends Handler
      * @return array
      * @throws \Exception
      */
-    public function dellUserData( int $userID ) : array
+    public function dellUserData( int $userID, string $userMTProd ) : array
     {
+        if ( !User::permission('nomUsers_edit') ) return ['error' => self::PERMISSION_DENIED ];
+
+        $access_old = (int)$this->findOne("SELECT access FROM users WHERE id='$userID'")['access'];
+        if ( $access_old === 1 && (User::getAccess() !== 1) ) return ['error' => self::PERMISSION_DENIED ];
+
         if ( !$this->checkID( $userID,'users' ) ) return ['error' => self::NO_SUCH_USER ];
+        
+        $this->baseSql("DELETE FROM user_permissions WHERE user_id='$userID'");
         if ($this->baseSql("DELETE FROM users WHERE id='$userID'") ) return ['success' => $userID];
 
         return ['error' => self::INSERT_UPDATE_FAIL ];
